tlf_handle_resolve: don't use finalized TLF IDs from iteam info
When chats got migrated to implicit teams, the current TLF ID got
written into the iteam sigchain.  KBFS uses that ID in the sigchain,
if it's available, in preparation for when we turn on KBFS iteams.
However, if the TLF gets finalized (because a user resets), the old ID
remains in the sigchain and KBFS will use it even when looking up the
non-finalized TLF ID.

We can fix this more permanently when we migrate existing KBFS TLFs to
iteams; at that point we can override the TLF ID using a new sigchain
link.

Issue: KBFS-3045